Individual Variable Effect

Individual Variable Effect

individual_variable_effect(x, ...)

# S3 method for explainer
individual_variable_effect(x, new_observation,
  method = "KernelSHAP", nsamples = 100, ...)

# S3 method for default
individual_variable_effect(x, data, predict_function,
  new_observation, label = class(x)[1], method = "KernelSHAP",
  nsamples = 100L, ...)

Arguments

x
a model to be explained, or an explainer created with function explain.

...
other parameters.

new_observation
an observation/observations to be explained. Required for local/instance level explainers. Columns in should correspond to columns in the data argument.

method
an estimation method of SHAP values. Currently the only availible is `KernelSHAP`.

nsamples
number of samples. Note that number must be as integer. Use `as.integer()`.

data
validation dataset. Used to determine univariate distributions, calculation of quantiles, correlations and so on. It will be extracted from `x` if it’s an explainer.

predict_function
predict function that operates on the model `x`. Since the model is a black box, the `predict_function` is the only interface to access values from the model. It should be a function that takes at least a model `x` and data and returns vector of predictions. If model response has more than a single number (like multiclass models) then this function should return a marix/data.frame of the size `m` x `d`, where `m` is the number of observations while `d` is the dimensionality of model response. It will be extracted from `x` if it’s an explainer.

label
name of the model. By default it’s extracted from the class attribute of the model

Value

an object of class individual_variable_effect with shap values of each variable for each new observation. Columns:

first d columns contains variable values.

_id_ - id of observation, number of row in `new_observation` data.

_ylevel_ - level of y

_yhat_ -predicted value for level of y

_yhat_mean_ - expected value of prediction, mean of all predictions

_vname_ - variable name

_attribution_ - attribution of variable

_sign_ a sign of attribution

_label_ a label of model

Examples

library("shapper")
library("DALEX")

#> Welcome to DALEX (version: 0.2.4).

library("randomForest")

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

#> 
#> Dołączanie pakietu: 'randomForest'

#> Następujący obiekt został zakryty z 'package:shapper':
#> 
#>     margin

Y_train <- HR$status
x_train <- HR[ , -6]
x_train$gender <- as.numeric(x_train$gender)
set.seed(123)
model_rf <- randomForest(x = x_train, y = Y_train)
p_fun <- function(x, data){
  predict(x, newdata = data, type = "prob")
  }
  res <- individual_variable_effect(x = model_rf, data = x_train,
                                    predict_function = p_fun,
                                    new_observation = x_train[1,])